divided main.rs
